+++ src/cchdo/hydro/checks.py
@@ -1,6 +1,14 @@
 from collections import defaultdict
 
+import numpy as np
 import xarray as xr
+
+from cchdo.hydro.exchange.exceptions import ExchangeDataFlagPairError
+from cchdo.hydro.exchange.flags import (
+    ExchangeBottleFlag,
+    ExchangeCTDFlag,
+    ExchangeSampleFlag,
+)
 
 
 def check_ancillary_variables(ds: xr.Dataset):
@@ -30,3 +38,90 @@ def check_ancillary_variables(ds: xr.Dataset):
 
     if errors := looks_ancillary - ancillary_variables_attrs.keys():
         raise ValueError(errors)
+
+
+def check_flags(dataset: xr.Dataset, raises=True):
+    """Check WOCE flag values agaisnt their param and ensure that the param either has a value or is "nan" depedning on the flag definition.
+
+    Return a boolean array of invalid locations?
+    """
+    woce_flags = {
+        "WOCESAMPLE": ExchangeBottleFlag,
+        "WOCECTD": ExchangeCTDFlag,
+        "WOCEBOTTLE": ExchangeSampleFlag,
+    }
+    flag_has_value = {
+        "WOCESAMPLE": {flag.value: flag.has_value for flag in ExchangeBottleFlag},
+        "WOCECTD": {flag.value: flag.has_value for flag in ExchangeCTDFlag},
+        "WOCEBOTTLE": {flag.value: flag.has_value for flag in ExchangeSampleFlag},
+    }
+    # In some cases, a coordinate variable might have flags, so we are not using filter_by_attrs
+    # get all the flag vars (that also have conventions)
+    flag_vars = []
+    for var_name in dataset.variables:
+        # do not replace the above with .items() it will give you xr.Variable objects that you don't want to use
+        # the following gets a real xr.DataArray
+        data = dataset[var_name]
+        if not {"standard_name", "conventions"} <= data.attrs.keys():
+            continue
+        if not any(flag in data.attrs["conventions"] for flag in woce_flags):
+            continue
+        if "status_flag" in data.attrs["standard_name"]:
+            flag_vars.append(var_name)
+
+    # match flags with their data vars
+    # it is legal in CF for one set of flags to apply to multiple vars
+    flag_errors = {}
+    for flag_var in flag_vars:
+        # get the flag and check attrs for defs
+        flag_da = dataset[flag_var]
+        conventions = None
+        for flag in woce_flags:
+            if flag_da.attrs.get("conventions", "").startswith(flag):
+                conventions = flag
+                break
+
+        # we don't know these flags, skip the check
+        if not conventions:
+            continue
+
+        allowed_values = np.array(list(flag_has_value[conventions]))
+        illegal_flags = ~flag_da.fillna(9).isin(allowed_values)
+        if np.any(illegal_flags):
+            illegal_flags.attrs["comments"] = (
+                f"This is a boolean array in the same shape as '{flag_da.name}' which is truthy where invalid values exist"
+            )
+            flag_errors[f"{flag_da.name}_value_errors"] = illegal_flags
+            continue
+
+        for var_name in dataset.variables:
+            data = dataset[var_name]
+            if "ancillary_variables" not in data.attrs:
+                continue
+            if flag_var not in data.attrs["ancillary_variables"].split(" "):
+                continue
+
+            # check data against flags
+            has_fill_f = [
+                flag
+                for flag, value in flag_has_value[conventions].items()
+                if value is False
+            ]
+
+            has_fill = flag_da.isin(has_fill_f) | np.isnan(flag_da)
+
+            # TODO deal with strs
+
+            if np.issubdtype(data.values.dtype, np.number):
+                fill_value_mismatch: xr.DataArray = ~(np.isfinite(data) ^ has_fill)  # type: ignore # numpy doesn't support __array_ufunc__ types yet
+                if np.any(fill_value_mismatch):
+                    fill_value_mismatch.attrs["comments"] = (
+                        f"This is a boolean array in the same shape as '{data.name}' which is truthy where invalid values exist"
+                    )
+                    flag_errors[f"{data.name}_value_errors"] = fill_value_mismatch
+
+    flag_errors_ds = xr.Dataset(flag_errors)
+    if raises and any(flag_errors_ds):
+        raise ExchangeDataFlagPairError(flag_errors_ds)
+
+    return flag_errors_ds